\name{BBEM-package}
\alias{BBEM-package}
\alias{BBEM}
\docType{package}
\title{
An Ensemble EM Algorithm for Bayesian Variable Selection
}
\description{
A simple EM algorithm and a Bayesian Bootstrap EM algorithm for variable selection, using continuous spike and slab prior.
}
\details{

}
\author{
Jin Wang

Maintainer: Jin Wang
}
\references{
http://arxiv.org/abs/1603.04360
}

